refactor(create-board):Board를 BoardFactory에서 생성하도록 수정
